+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- sendNotification method on portal service
+
 ## 1.0.0
 ### Changed
 - swapped the deprecated `ember-network` for `ember-fetch`